refactored analytics, introduced vero and managing anonymous ids (#2997)
* refactored analytics, introduced vero and managing anonymous ids

* Rename DNT var

* Remove unused loadTour

* Handle pageviews

* Also sign in state when loading app
