+++ packages/app/src/app/overmind/effects/index.ts
@@ -5,7 +5,6 @@ export { default as browser } from './browser';
 export { default as connection } from './connection';
 export { default as fsSync } from './fsSync';
 export { default as jsZip } from './jsZip';
-export { default as jwt } from './jwt';
 export { default as keybindingManager } from './keybindingManager';
 export { default as live } from './live';
 export { default as moduleRecover } from './moduleRecover';
@@ -25,4 +24,4 @@ export { default as codesandboxApi } from './codesandboxApi';
 export { default as themes } from './themes';
 export { default as executor } from './executor';
 export { default as stripe } from './stripe';
-export { default as chameleon } from './chameleon';
+export { default as jwt } from './jwt';
