+++ packages/app/src/app/overmind/factories.ts
@@ -1,5 +1,6 @@
 import { Contributor } from '@codesandbox/common/lib/types';
-import { json, IState, IDerive } from 'overmind';
+import { IDerive, IState, json } from 'overmind';
+
 import { AsyncAction } from '.';
 
 export const withLoadApp = <T>(
@@ -30,6 +31,8 @@ export const withLoadApp = <T>(
       state.user = await effects.api.getCurrentUser();
       actions.internal.setPatronPrice();
       actions.internal.setSignedInCookie();
+      effects.analytics.identify('signed_in', true);
+      effects.analytics.setUserId(state.user.id);
       actions.internal.showUserSurveyIfNeeded();
       effects.live.connect();
       actions.userNotifications.internal.initialize();
@@ -39,9 +42,11 @@ export const withLoadApp = <T>(
         'Your session seems to be expired, please log in again...'
       );
       effects.jwt.reset();
+      effects.analytics.setAnonymousId();
     }
   } else {
     effects.jwt.reset();
+    effects.analytics.setAnonymousId();
   }
 
   if (continueAction) {
