+++ packages/app/src/app/pages/index.tsx
@@ -4,7 +4,7 @@ import { DragDropContext } from 'react-dnd';
 import _debug from '@codesandbox/common/lib/utils/debug';
 import { Toasts, NotificationStatus } from '@codesandbox/notifications';
 import { notificationState } from '@codesandbox/common/lib/utils/notifications';
-import send, { DNT } from '@codesandbox/common/lib/utils/analytics';
+import { DNT, trackPageview } from '@codesandbox/common/lib/utils/analytics';
 import theme from '@codesandbox/common/lib/theme';
 import { Button } from '@codesandbox/common/lib/components/Button';
 import Loadable from 'app/utils/Loadable';
@@ -79,16 +79,10 @@ const RoutesComponent: React.FC = () => {
         render={({ location }) => {
           if (process.env.NODE_ENV === 'production') {
             routeDebugger(
-              `Sending '${location.pathname + location.search}' to ga.`
+              `Sending '${location.pathname + location.search}' to analytics.`
             );
-            if (typeof (window as any).ga === 'function' && !DNT) {
-              (window as any).ga(
-                'set',
-                'page',
-                location.pathname + location.search
-              );
-
-              send('pageview', { path: location.pathname + location.search });
+            if (!DNT) {
+              trackPageview();
             }
           }
           return null;
